add require 'magicbell' to magicbell-rails.rb (#7)
When running this from nexl360:
`Magicbell::Rails.fetch_categories(external_id:)`

I get the following error:
`uninitialized constant Magicbell::Rails::MagicBell (NameError)`

This is because 'magicbell' wasn't required in the class. This didn't
get picked up on in the specs because we require it in the spec/config
so no error was getting thrown.