add create and user_exists? to new User module (#8)
### Summary
We need to check if a user exists before fetching their categories. If
we don't do this we end up with a `MagicBell::Client::HTTPError` because
the external_id we searched for doesn't exist. We also need to add a
create in the case that the user doesn't exist and we need to create one
in MagicBell ourselves.

### Changes
- Added `User.create`
- Added `User.exists?` (I had to use .notification_preferences.retrieve
because magicbell does not have an endpoint for looking up user by
external_id or email)
- Updated version from `0.3.1` -> `0.4.0`